requirements.txt
# fastapi
fastapi==0.95.0
uvicorn==0.21.1

# apis
openai==0.27.4
google-cloud-translate==3.11.1

# for testing
pytest==7.3.0
httpx==0.24.0
python-dotenv==1.0.0
